Extract export dependency walker into its own module.
DependencyWalkResult and walk_dependency_tree move to export_dependency_walk.py
so export_context.py stays focused on probe/query orchestration.
