+++ packages/api/api/analytics/export_context.py
@@ -6,10 +6,9 @@ from collections.abc import Callable, Mapping
 from dataclasses import dataclass, field
 from typing import Any
 
+from api.analytics.export_dependency_walk import walk_dependency_tree
 from api.analytics.export_errors import ExportCycleDetectedError
 from api.analytics.export_types import (
-    EnsureDependency,
-    EnsureMissingStep,
     ExportProbeResult,
     ExportQueryResult,
     ExportScope,
@@ -24,15 +23,6 @@ from api.analytics.options import TurnAnalyticsOptions
 from api.models.game import TurnInfo
 
 INLINE_ENSURE_MAX_MISSING_STEPS = 5
-
-
-@dataclass
-class DependencyWalkResult:
-    turn_unavailable: UnavailableReason | None = None
-    missing_steps: list[EnsureMissingStep] = field(default_factory=list)
-    pending_ensure: list[tuple[str, ExportScope, AnalyticExportCatalog]] = field(
-        default_factory=list,
-    )
 
 
 @dataclass(frozen=True)
@@ -76,7 +66,8 @@ class AnalyticQueryContext:
         prep = self._prepare_export_request(analytic_id, scope_overrides)
         if not isinstance(prep, PreparedExportRequest):
             return self._probe_unavailable(prep)
-        walk_result = self._walk_dependency_tree(
+        walk_result = walk_dependency_tree(
+            self,
             analytic_id,
             prep.scope,
             visiting=set(),
@@ -127,7 +118,8 @@ class AnalyticQueryContext:
             self._memo[resolution_key] = result
             return result
 
-        walk_result = self._walk_dependency_tree(
+        walk_result = walk_dependency_tree(
+            self,
             analytic_id,
             scope,
             visiting=set(),
@@ -141,11 +133,7 @@ class AnalyticQueryContext:
 
         total_missing = len(walk_result.missing_steps)
         blocked_inline = total_missing > INLINE_ENSURE_MAX_MISSING_STEPS
-        if (
-            blocked_inline
-            and not force_inline_ensure
-            and self.enforce_inline_ensure_threshold
-        ):
+        if blocked_inline and not force_inline_ensure and self.enforce_inline_ensure_threshold:
             result = self._unavailable("ensure_blocked")
             self._memo[resolution_key] = result
             return result
@@ -217,99 +205,6 @@ class AnalyticQueryContext:
                 return "invalid_scope"
         return None
 
-    def _dependency_needs_processing(
-        self,
-        analytic_id: str,
-        scope: ExportScope,
-        catalog: AnalyticExportCatalog | None,
-    ) -> bool:
-        if catalog is None or catalog.is_empty:
-            return False
-        if self._is_at_baseline(analytic_id, scope, catalog):
-            return False
-        if self._is_persisted(analytic_id, scope, catalog):
-            return False
-        return True
-
-    def _walk_dependency_tree(
-        self,
-        analytic_id: str,
-        scope: ExportScope,
-        *,
-        visiting: set[tuple[str, ExportScope]],
-        check_turn_availability: bool,
-        detect_ensure_cycles: bool,
-    ) -> DependencyWalkResult:
-        result = DependencyWalkResult()
-        visit_key = (analytic_id, scope)
-        if visit_key in visiting:
-            if detect_ensure_cycles:
-                raise ExportCycleDetectedError(
-                    f"Analytic export ensure cycle detected for {analytic_id!r} "
-                    f"at turn {scope.turn} with player_id {scope.player_id!r}"
-                )
-            return result
-
-        visiting.add(visit_key)
-        try:
-            catalog = self.export_registry.get(analytic_id)
-            if not self._dependency_needs_processing(analytic_id, scope, catalog):
-                return result
-
-            assert catalog is not None
-
-            for dependency in catalog.ensure_dependencies:
-                dependency_scope = self._dependency_scope(scope, dependency)
-                if dependency_scope.turn < 1:
-                    continue
-
-                if self.load_turn(dependency_scope.turn) is None:
-                    if check_turn_availability:
-                        result.turn_unavailable = "turn_not_stored"
-                        return result
-                    result.missing_steps.append(
-                        EnsureMissingStep(
-                            analytic_id=dependency.analytic_id,
-                            turn=dependency_scope.turn,
-                            player_id=dependency_scope.player_id,
-                            status="not_persisted",
-                        )
-                    )
-                    continue
-
-                dependency_catalog = self.export_registry.get(dependency.analytic_id)
-                if check_turn_availability and (
-                    dependency_catalog is None or dependency_catalog.is_empty
-                ):
-                    continue
-
-                nested = self._walk_dependency_tree(
-                    dependency.analytic_id,
-                    dependency_scope,
-                    visiting=visiting,
-                    check_turn_availability=check_turn_availability,
-                    detect_ensure_cycles=detect_ensure_cycles,
-                )
-                if nested.turn_unavailable is not None:
-                    result.turn_unavailable = nested.turn_unavailable
-                    return result
-
-                result.missing_steps.extend(nested.missing_steps)
-                result.pending_ensure.extend(nested.pending_ensure)
-
-            result.missing_steps.append(
-                EnsureMissingStep(
-                    analytic_id=analytic_id,
-                    turn=scope.turn,
-                    player_id=scope.player_id,
-                    status="not_persisted",
-                )
-            )
-            result.pending_ensure.append((analytic_id, scope, catalog))
-            return result
-        finally:
-            visiting.discard(visit_key)
-
     def _apply_pending_ensure(
         self,
         pending_ensure: list[tuple[str, ExportScope, AnalyticExportCatalog]],
@@ -318,49 +213,6 @@ class AnalyticQueryContext:
             if catalog.ensure_export is not None:
                 catalog.ensure_export(self, scope)
             self._ensured_scopes.add((analytic_id, scope))
-
-    def _dependency_scope(
-        self,
-        scope: ExportScope,
-        dependency: EnsureDependency,
-    ) -> ExportScope:
-        player_id = scope.player_id
-        if dependency.player_id != "same":
-            player_id = None
-        return ExportScope(
-            game_id=scope.game_id,
-            perspective=scope.perspective,
-            turn=scope.turn + dependency.turn_delta,
-            player_id=player_id,
-        )
-
-    def _is_at_baseline(
-        self,
-        analytic_id: str,
-        scope: ExportScope,
-        catalog: AnalyticExportCatalog,
-    ) -> bool:
-        if scope.turn <= 1 and not catalog.ensure_dependencies:
-            return True
-        if scope.turn <= 1:
-            for dependency in catalog.ensure_dependencies:
-                dependency_scope = self._dependency_scope(scope, dependency)
-                if dependency_scope.turn < 1:
-                    return True
-        return False
-
-    def _is_persisted(
-        self,
-        analytic_id: str,
-        scope: ExportScope,
-        catalog: AnalyticExportCatalog,
-    ) -> bool:
-        scope_key = (analytic_id, scope)
-        if scope_key in self._ensured_scopes:
-            return True
-        if catalog.is_persisted is None:
-            return False
-        return catalog.is_persisted(self, scope)
 
     def _materialize_tree(
         self,
